+++ search/serializers.py
@@ -2,6 +2,8 @@ import markdown2
 from bs4 import BeautifulSoup
 from urllib3.util import parse_url
 
+from django.utils.text import Truncator
+
 from directory_constants import urls
 
 
@@ -19,6 +21,27 @@ def parse_search_results(content):
             url = parse_url(result['url'])
             result['url'] = urls.SERVICES_EVENTS + url.request_uri
 
+    def abridge_long_contents(result):
+        if 'content' in result:
+            result['content'] = Truncator(result['content']).chars(160)
+
+    def format_display_type(result):
+        mappings = {
+            'dit:Event': 'Event',
+            'Event': 'Event',
+            'dit:Opportunity': 'Export opportunity',
+            'Opportunity': 'Export opportunity',
+            'Market': 'Online marketplace',
+            'dit:Market': 'Online marketplace',
+            'Article': 'Article',
+            'dit:Article': 'Article',
+            'Service': 'Service',
+            'dit:Service': 'Service'
+        }
+        for value, replacement in mappings.items():
+            if value in result['type']:
+                result['type'] = replacement
+
     results = [hit['_source'] for hit in content['hits']['hits']]
 
     # This removes HTML tags and markdown received from CMS results
@@ -29,10 +52,7 @@ def parse_search_results(content):
     for result in results:
         strip_html(result)
         format_events_url(result)
-
-    # Abridge long text snippets
-    for result in results:
-        if ('content' in result) and (len(result['content']) > 160):
-            result['content'] = result['content'][0:160] + '...'
+        format_display_type(result)
+        abridge_long_contents(result)
 
     return results